+++ docs/source/user_guide/changelog.rst
@@ -5,6 +5,8 @@ Changelog
 Authors since :ref:`5.0.0`: Jordi Torrents
 
 - Merge network architectures for crossings and identification.
+- Add ``--size`` parameter for individual :ref:`video generators`.
+
 
 5.2.1
 =====